+++ npm/cmd/debug.go
