+++ npm/config/config_linux.go
@@ -3,7 +3,7 @@
 
 package npmconfig
 
-// TODO: solidify config paths
+//GetConfigPath TODO: solidify config paths
 func GetConfigPath() string {
 	return "/etc/azure/azure-vnet/azure-npm.json"
 }
