+++ npm/config/config_windows.go
@@ -3,7 +3,7 @@
 
 package npmconfig
 
-//TODO: solidify config paths
+//GetConfigPath TODO: solidify config paths
 func GetConfigPath() string {
 	return "c:\\k\\azure-npm\\azure-npm.json"
 }
